+++ Rx.playground/Pages/Combining_Operators.xcplaygroundpage/Contents.swift
@@ -155,6 +155,36 @@ example("switchLatest") {
 }
 /*:
  > In this example, adding ⚾️ onto `subject1` after adding `subject2` to `subjectsSubject` has no effect, because only the most recent inner `Observable` sequence (`subject2`) will emit elements.
+
+ ----
+ ## `withLatestFrom`
+ Merges two observable sequences into one observable sequence by combining each element from the first source with the latest element from the second source, if any.
+ */
+example("withLatestFrom") {
+    let disposeBag = DisposeBag()
+    
+    let foodSubject = PublishSubject<String>()
+    let drinksSubject = PublishSubject<String>()
+    
+    foodSubject.asObservable()
+        .withLatestFrom(drinksSubject) { "\($0) + \($1)" }
+        .subscribe(onNext: { print($0) })
+        .disposed(by: disposeBag)
+    
+    foodSubject.onNext("🥗")
+    
+    drinksSubject.onNext("☕️")
+    foodSubject.onNext("🥐")
+    
+    drinksSubject.onNext("🍷")
+    foodSubject.onNext("🍔")
+    
+    foodSubject.onNext("🍟")
+    
+    drinksSubject.onNext("🍾")
+}
+/*:
+ > In this example 🥗 is not printed because `drinksSubject` did not emit any values before 🥗 was received. The last drink (🍾) will be printed whenever `foodSubject` will emit another event.
  */
 
 //: [Next](@next) - [Table of Contents](Table_of_Contents)