+++ RxCocoa/Common/Observable+Bind.swift
@@ -61,7 +61,31 @@ extension ObservableType {
     public func bind<R1, R2>(to binder: (Self) -> (R1) -> R2, curriedArgument: R1) -> R2 {
         binder(self)(curriedArgument)
     }
+    
+    /**
+    Subscribes an element handler to an observable sequence.
+    In case error occurs in debug mode, `fatalError` will be raised.
+    In case error occurs in release mode, `error` will be logged.
 
+     - Note: If `object` can't be retained, none of the other closures will be invoked.
+     
+    - parameter object: The object to provide an unretained reference on.
+    - parameter onNext: Action to invoke for each element in the observable sequence.
+    - returns: Subscription object used to unsubscribe from the observable sequence.
+    */
+    public func bind<Object: AnyObject>(
+        with object: Object,
+        onNext: @escaping (Object, Element) -> Void
+    ) -> Disposable {
+        self.subscribe(onNext: { [weak object] in
+            guard let object = object else { return }
+            onNext(object, $0)
+        },
+        onError: { error in
+            rxFatalErrorInDebug("Binding error: \(error)")
+        })
+    }
+    
     /**
     Subscribes an element handler to an observable sequence.
     In case error occurs in debug mode, `fatalError` will be raised.