+++ RxCocoa/Traits/Driver/Driver+Subscription.swift
@@ -19,7 +19,7 @@ extension SharedSequenceConvertibleType where SharingStrategy == DriverSharingSt
 
     In this form it's equivalent to `subscribe` method, but it communicates intent better.
 
-    - parameter observer: Observer that receives events.
+    - parameter observers: Observers that receives events.
     - returns: Disposable object that can be used to unsubscribe the observer from the subject.
     */
     public func drive<Observer: ObserverType>(_ observers: Observer...) -> Disposable where Observer.Element == Element {
@@ -37,7 +37,7 @@ extension SharedSequenceConvertibleType where SharingStrategy == DriverSharingSt
 
      In this form it's equivalent to `subscribe` method, but it communicates intent better.
 
-     - parameter observer: Observer that receives events.
+     - parameter observers: Observers that receives events.
      - returns: Disposable object that can be used to unsubscribe the observer from the subject.
      */
     public func drive<Observer: ObserverType>(_ observers: Observer...) -> Disposable where Observer.Element == Element? {
@@ -54,7 +54,7 @@ extension SharedSequenceConvertibleType where SharingStrategy == DriverSharingSt
     Creates new subscription and sends elements to `BehaviorRelay`.
     This method can be only called from `MainThread`.
 
-    - parameter relay: Target relay for sequence elements.
+    - parameter relays: Target relays for sequence elements.
     - returns: Disposable object that can be used to unsubscribe the observer from the relay.
     */
     public func drive(_ relays: BehaviorRelay<Element>...) -> Disposable {
@@ -68,7 +68,7 @@ extension SharedSequenceConvertibleType where SharingStrategy == DriverSharingSt
      Creates new subscription and sends elements to `BehaviorRelay`.
      This method can be only called from `MainThread`.
 
-     - parameter relay: Target relay for sequence elements.
+     - parameter relays: Target relays for sequence elements.
      - returns: Disposable object that can be used to unsubscribe the observer from the relay.
      */
     public func drive(_ relays: BehaviorRelay<Element?>...) -> Disposable {
@@ -82,7 +82,7 @@ extension SharedSequenceConvertibleType where SharingStrategy == DriverSharingSt
     Creates new subscription and sends elements to `ReplayRelay`.
     This method can be only called from `MainThread`.
 
-    - parameter relay: Target relay for sequence elements.
+    - parameter relays: Target relays for sequence elements.
     - returns: Disposable object that can be used to unsubscribe the observer from the relay.
     */
     public func drive(_ relays: ReplayRelay<Element>...) -> Disposable {
@@ -96,7 +96,7 @@ extension SharedSequenceConvertibleType where SharingStrategy == DriverSharingSt
      Creates new subscription and sends elements to `ReplayRelay`.
      This method can be only called from `MainThread`.
 
-     - parameter relay: Target relay for sequence elements.
+     - parameter relays: Target relays for sequence elements.
      - returns: Disposable object that can be used to unsubscribe the observer from the relay.
      */
     public func drive(_ relays: ReplayRelay<Element?>...) -> Disposable {
@@ -140,6 +140,34 @@ extension SharedSequenceConvertibleType where SharingStrategy == DriverSharingSt
     /**
     Subscribes an element handler, a completion handler and disposed handler to an observable sequence.
     This method can be only called from `MainThread`.
+     
+    Also, take in an object and provide an unretained, safe to use (i.e. not implicitly unwrapped), reference to it along with the events emitted by the sequence.
+    
+    Error callback is not exposed because `Driver` can't error out.
+     
+     - Note: If `object` can't be retained, none of the other closures will be invoked.
+     
+    - parameter object: The object to provide an unretained reference on.
+    - parameter onNext: Action to invoke for each element in the observable sequence.
+    - parameter onCompleted: Action to invoke upon graceful termination of the observable sequence.
+    gracefully completed, errored, or if the generation is canceled by disposing subscription)
+    - parameter onDisposed: Action to invoke upon any type of termination of sequence (if the sequence has
+    gracefully completed, errored, or if the generation is canceled by disposing subscription)
+    - returns: Subscription object used to unsubscribe from the observable sequence.
+    */
+    public func drive<Object: AnyObject>(
+        with object: Object,
+        onNext: ((Object, Element) -> Void)? = nil,
+        onCompleted: ((Object) -> Void)? = nil,
+        onDisposed: ((Object) -> Void)? = nil
+    ) -> Disposable {
+        MainScheduler.ensureRunningOnMainThread(errorMessage: errorMessage)
+        return self.asObservable().subscribe(with: object, onNext: onNext, onCompleted: onCompleted, onDisposed: onDisposed)
+    }
+    
+    /**
+    Subscribes an element handler, a completion handler and disposed handler to an observable sequence.
+    This method can be only called from `MainThread`.
     
     Error callback is not exposed because `Driver` can't error out.
     
@@ -150,7 +178,11 @@ extension SharedSequenceConvertibleType where SharingStrategy == DriverSharingSt
     gracefully completed, errored, or if the generation is canceled by disposing subscription)
     - returns: Subscription object used to unsubscribe from the observable sequence.
     */
-    public func drive(onNext: ((Element) -> Void)? = nil, onCompleted: (() -> Void)? = nil, onDisposed: (() -> Void)? = nil) -> Disposable {
+    public func drive(
+        onNext: ((Element) -> Void)? = nil,
+        onCompleted: (() -> Void)? = nil,
+        onDisposed: (() -> Void)? = nil
+    ) -> Disposable {
         MainScheduler.ensureRunningOnMainThread(errorMessage: errorMessage)
         return self.asObservable().subscribe(onNext: onNext, onCompleted: onCompleted, onDisposed: onDisposed)
     }