+++ Sources/AllTestz/main.swift
@@ -274,6 +274,9 @@ final class DriverTest_ : DriverTest, RxTestCase {
     ("testDriveOptionalReplayRelay2", DriverTest.testDriveOptionalReplayRelay2),
     ("testDriveReplayRelays2", DriverTest.testDriveReplayRelays2),
     ("testDriveReplayRelayNoAmbiguity", DriverTest.testDriveReplayRelayNoAmbiguity),
+    ("testDriveWithNext", DriverTest.testDriveWithNext),
+    ("testDriveWithError", DriverTest.testDriveWithError),
+    ("testDriveWithCompleted", DriverTest.testDriveWithCompleted),
     ] }
 }
 
@@ -325,6 +328,8 @@ final class InfallibleTest_ : InfallibleTest, RxTestCase {
     ("testAnonymousInfallible_detachesOnDispose", InfallibleTest.testAnonymousInfallible_detachesOnDispose),
     ("testAnonymousInfallible_detachesOnComplete", InfallibleTest.testAnonymousInfallible_detachesOnComplete),
     ("testAsInfallible_never", InfallibleTest.testAsInfallible_never),
+    ("testSubscribeWithNext", InfallibleTest.testSubscribeWithNext),
+    ("testSubscribeWithError", InfallibleTest.testSubscribeWithError),
     ] }
 }
 
@@ -1629,6 +1634,7 @@ final class ObservableTest_ : ObservableTest, RxTestCase {
     ("testAsObservable_asObservable", ObservableTest.testAsObservable_asObservable),
     ("testAsObservable_hides", ObservableTest.testAsObservable_hides),
     ("testAsObservable_never", ObservableTest.testAsObservable_never),
+    ("testSubscribeWithNext", ObservableTest.testSubscribeWithNext),
     ] }
 }
 
@@ -1891,6 +1897,9 @@ final class ReactiveTests_ : ReactiveTests, RxTestCase {
 
     static var allTests: [(String, (ReactiveTests_) -> () -> Void)] { return [
     ("testEnablesMutations", ReactiveTests.testEnablesMutations),
+    ("testReactiveStruct", ReactiveTests.testReactiveStruct),
+    ("testReactiveProtocol", ReactiveTests.testReactiveProtocol),
+    ("testDynamicLookup", ReactiveTests.testDynamicLookup),
     ] }
 }
 
@@ -2052,6 +2061,9 @@ final class SignalTests_ : SignalTests, RxTestCase {
     ("testEmitOptionalReplayRelay2", SignalTests.testEmitOptionalReplayRelay2),
     ("testEmitReplayRelays2", SignalTests.testEmitReplayRelays2),
     ("testEmitReplayRelayNoAmbiguity", SignalTests.testEmitReplayRelayNoAmbiguity),
+    ("testEmitWithNext", SignalTests.testEmitWithNext),
+    ("testEmitWithError", SignalTests.testEmitWithError),
+    ("testEmitWithCompleted", SignalTests.testEmitWithCompleted),
     ] }
 }
 
@@ -2144,6 +2156,21 @@ final class VirtualSchedulerTest_ : VirtualSchedulerTest, RxTestCase {
     ("testVirtualScheduler_stop", VirtualSchedulerTest.testVirtualScheduler_stop),
     ("testVirtualScheduler_sleep", VirtualSchedulerTest.testVirtualScheduler_sleep),
     ("testVirtualScheduler_stress", VirtualSchedulerTest.testVirtualScheduler_stress),
+    ] }
+}
+
+final class WithUnretainedTests_ : WithUnretainedTests, RxTestCase {
+    #if os(macOS)
+    required override init() {
+        super.init()
+    }
+    #endif
+
+    static var allTests: [(String, (WithUnretainedTests_) -> () -> Void)] { return [
+    ("testObjectAttached", WithUnretainedTests.testObjectAttached),
+    ("testObjectDeallocates", WithUnretainedTests.testObjectDeallocates),
+    ("testObjectDeallocatesSequenceCompletes", WithUnretainedTests.testObjectDeallocatesSequenceCompletes),
+    ("testResultsSelector", WithUnretainedTests.testResultsSelector),
     ] }
 }
 #if os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
@@ -2262,5 +2289,6 @@ func XCTMain(_ tests: [() -> Void]) {
         testCase(SignalTests_.allTests),
         testCase(SingleTest_.allTests),
         testCase(VirtualSchedulerTest_.allTests),
+        testCase(WithUnretainedTests_.allTests),
     ])
 //}